Created seperate heartrate files and fixed usart declarations
