+++ heartrate.h
@@ -1,0 +1,4 @@
+extern uint8_t smooth_signal(uint8_t);
+extern void heartbeatCalc(char *);
+extern void timer1_init();
+extern unsigned long millis();
